# spanlogger: skip formatting if trace is not sampled